fixed unautherized bug
